backend/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-jose==3.3.0
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyjwt==2.8.0 